Big rework of load/save.  Saves weights PLUS network architecture, and better.
Things changed:

Layers are now referred to by names rather than class references so that
they can be deserialized unambiguously while still being stored in tiny
files.

Option to construct any object with filename is removed as that is
confusing and not needed since we now have everything bound together
with the NeuralCIs object.

Sequential object is wrapped up in _SequentialNet that can save all the
details of its own architecture.  This is also refactored to mean that
if we change later to not have the same layer type at each hidden layer,
this will automatically be handled without making old saved files obsolete.

The name that is saved is now the name of the folder which contains all
the files and is also relative to cwd, rather than relative to
cwd/savedmodels.  This will make moving saved nets around and renaming
them much more straightforward.

_SimulatorNet.trainable_weights is no longer stored in the object but
instead computed live.  This means that reloading the backend doesn't
leave us all stale.