Increase the amount of sampling we do outside of the valid estimates box.
There was a problem where the p-values were coming out a little too high
on one side and a little too low on the other.  Just tried once so far
but on that one try, this switch made it become centred again.  (Perhaps
just a bit too noisy, and only revealing enough params on one side?)