Add "smoothers" to _SamplingFeelerNet, for adding noise to pre-cached data
For each pre-cached datapoint, the Cholesky factor of its sampling matrix
is also stored, and this is used to add noise to the cached datapoints.

Currently amount of noise to be added is estimated from number of points
in the sampling rectangle vs the sum of their Cholesky determinants.  This
currently targets the ratio of 1, but we might want to increase that for
more smoothing.  Further experiments to follow.