Now we have smoothing in the feeler cache, we can track the training loss
Previously, the feeler net was the one net with a fixed dataset, and
allowing it to choose its fitting based on training loss would give
horrible overfit to that dataset.  However, now that we add enough
noise that there isn't the same opportunity to overfit, we can track
the training loss instead, which will allow a tighter fit, and also
reduce the chances that we end up overfitting a validation set (by
constantly only making small steps that favour the validation set).