Move discretization of sample sizes back to param sampling training.  Needs huge refactor.
Main content of refactor is to make neuralcis store all the variable
distributions in the order of the net, rather than of the simulation,
which then makes the integration of all the various new functions
pretty clean.  This includes in particular the new preprocess net
interface needed for this change, and the ability to make it work only
for known params.