Add quadratic approximation to try to find whether a param sample is in outer region.
In the end, this does not work well, because it assumes that the route
from the param sample to the estimates box will be along a straight line,
where it will generally tend to be a step to the side (e.g. shift mu,
keep sigma large) and then a step towards the centre.  Abandon for now.