Tighten / relax some thresholds.  Makes biserial work and welch still works.
Need to find more principled approaches to these problems.  With the
IS_INSIDE_NET_THRESHOLD at 0.1, the biserial ended up with a hole in
the middle of sampling space.  And with tolerances too low, it kept having
to restore old checkpoints, even though it was training just fine.

(Perhaps don't even need the tolerances any more?)